Clang format + macOS compilation errors
Inspect and clean up the util library


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error

+++ src/genie/core/format-exporter-compressed.h
@@ -40,7 +40,7 @@ class FormatExporterCompressed : public util::Drain<AccessUnit> {
     /**
      *
      */
-    void flushIn(size_t& pos) override;
+    void flushIn(uint64_t& pos) override;
 
     /**
      *
